+++ test/with_traefik/asynchronous/test_svn.py
@@ -48,7 +48,7 @@ class TestSvnCapability(TraefikTestCase):
         async with AsyncHTTPSConnectionPool(
             self.host,
             self.https_port,
-            timeout=1,
+            timeout=5,
             retries=0,
             ca_certs=self.ca_authority,
             resolver=self.test_async_resolver,
@@ -62,7 +62,7 @@ class TestSvnCapability(TraefikTestCase):
         async with AsyncHTTPSConnectionPool(
             self.host,
             self.https_port,
-            timeout=1,
+            timeout=5,
             retries=0,
             ca_certs=self.ca_authority,
             disabled_svn={HttpVersion.h3},
@@ -76,7 +76,7 @@ class TestSvnCapability(TraefikTestCase):
         async with AsyncHTTPSConnectionPool(
             self.host,
             self.https_port,
-            timeout=1,
+            timeout=5,
             retries=0,
             ca_certs=self.ca_authority,
             disabled_svn={HttpVersion.h2},
@@ -93,7 +93,7 @@ class TestSvnCapability(TraefikTestCase):
         p = AsyncHTTPSConnectionPool(
             self.host,
             self.https_port,
-            timeout=1,
+            timeout=5,
             retries=0,
             ca_certs=self.ca_authority,
             disabled_svn={HttpVersion.h11},
@@ -112,7 +112,7 @@ class TestSvnCapability(TraefikTestCase):
             p = AsyncHTTPSConnectionPool(
                 self.host,
                 self.https_port,
-                timeout=1,
+                timeout=5,
                 retries=0,
                 ca_certs=self.ca_authority,
                 disabled_svn={HttpVersion.h11, HttpVersion.h2, HttpVersion.h3},
@@ -125,7 +125,7 @@ class TestSvnCapability(TraefikTestCase):
             p = AsyncHTTPConnectionPool(  # type: ignore[assignment]
                 self.host,
                 self.http_port,
-                timeout=1,
+                timeout=5,
                 retries=0,
                 disabled_svn={HttpVersion.h11, HttpVersion.h2},
                 resolver=self.test_async_resolver,
@@ -137,7 +137,7 @@ class TestSvnCapability(TraefikTestCase):
         async with AsyncHTTPSConnectionPool(
             self.host,
             self.https_alt_port,
-            timeout=1,
+            timeout=5,
             retries=False,
             ca_certs=self.ca_authority,
             resolver=self.test_async_resolver,
@@ -155,11 +155,40 @@ class TestSvnCapability(TraefikTestCase):
                     == 'h3-25=":443"; ma=3600, h2=":443"; ma=3600'
                 )
 
+    @pytest.mark.usefixtures("requires_http3")
+    async def test_misleading_upgrade_h3(self) -> None:
+        dumb_cache: dict[tuple[str, int], tuple[str, int] | None] = dict()
+
+        async with AsyncHTTPSConnectionPool(
+            self.host,
+            self.https_alt_port,
+            retries=False,
+            ca_certs=self.ca_authority,
+            resolver=self.test_async_resolver,
+            preemptive_quic_cache=dumb_cache,
+        ) as p:
+            for i in range(2):
+                resp = await p.request(
+                    "GET",
+                    "/response-headers?Alt-Svc=h3-25%3D%22%3A443%22%3B%20ma%3D3600%2C%20h3%3D%22%3A6547%22%3B%20ma%3D3600",
+                )
+
+                assert resp.version == 20
+                assert "Alt-Svc" in resp.headers
+                assert (
+                    resp.headers.get("Alt-Svc")
+                    == 'h3-25=":443"; ma=3600, h3=":6547"; ma=3600'
+                )
+
+        # this asserts tell us that we correctly unset the entry from "preemptive_quic_cache"
+        # because the alt-svc entry was misleading, thus invalid.
+        assert len(dumb_cache) == 0
+
     async def test_illegal_upgrade_h3(self) -> None:
         async with AsyncHTTPSConnectionPool(
             self.host,
             self.https_alt_port,
-            timeout=1,
+            timeout=5,
             retries=False,
             ca_certs=self.ca_authority,
             resolver=self.test_async_resolver,
@@ -182,7 +211,7 @@ class TestSvnCapability(TraefikTestCase):
         async with AsyncHTTPSConnectionPool(
             self.host,
             self.https_alt_port,
-            timeout=1,
+            timeout=5,
             retries=False,
             ca_certs=self.ca_authority,
             resolver=self.test_async_resolver,
@@ -204,7 +233,7 @@ class TestSvnCapability(TraefikTestCase):
         async with AsyncHTTPSConnectionPool(
             self.host,
             self.https_alt_port,
-            timeout=1,
+            timeout=5,
             retries=False,
             ca_certs=self.ca_authority,
             resolver=self.test_async_resolver,
@@ -334,7 +363,7 @@ class TestSvnCapability(TraefikTestCase):
         async with AsyncHTTPConnectionPool(
             self.host,
             self.http_alt_port,
-            timeout=1,
+            timeout=5,
             retries=False,
             resolver=self.test_async_resolver,
         ) as p: